+++ src/projectgoldstarsx/ErrorLogSettings.java
@@ -19,7 +19,7 @@ public class ErrorLogSettings
     private void errorLogSettings()
     {
         settingsFrame = new ProgramWindow("Error Log Settings");
-        settingsFrame.setSize(850 * ProjectGoldStarsX.multiplier, 450 * ProjectGoldStarsX.multiplier);
+        settingsFrame.setSize(800 * ProjectGoldStarsX.multiplier, 450 * ProjectGoldStarsX.multiplier);
         settingsFrame.setLayout(new GridLayout(3, 1));
         settingsFrame.setStandardMenuBar();
         settingsFrame.add(Components.headerLabel("Error Log Search Case Sensitivity"));